requirements.txt
annotated-types==0.5.0
anyio==3.6.2
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.0
exceptiongroup==1.1.1
fastapi==0.95.2
filelock==3.12.2
fsspec==2023.6.0
h11==0.14.0
huggingface-hub==0.14.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.2.1
mypy==1.4.1
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.3
outcome==1.2.0
packaging==23.1
pandas==2.0.2
pycparser==2.21
pydantic==1.10.7
pydantic-core==2.4.0
PySocks==1.7.1
pymongo==4.11
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
selenium==4.10.0
setuptools==67.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
starlette==0.27.0
sympy==1.11.1
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
tqdm==4.65.0
transformers==4.29.2
trio==0.22.0
trio-websocket==0.10.2
types-beautifulsoup4==4.12.0.1
types-pytz==2023.3.0.0
typing-extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.22.0
webdriver-manager==3.8.6
websocket-client==1.5.1
wsproto==1.2.0
sqlalchemy
pymysql
cryptography